hash-gen/requirements.txt
click==8.1.7
cloudpathlib[gs]
#google-api-core==2.14.0
google-auth==2.25.1
#google-cloud-core==2.3.3
google-cloud-storage==2.13.0
ImageHash==4.3.1
Pillow==10.1.0
dataset
psycopg2_binary
sqlalchemy-bigquery[bqstorage]